Issue #4: Reload Catapult refactored
The sequence of the reload catapult completely rewritten to make the
logic more transparent. Hopefully will fix the issue.

+++ src/Subsystems/Catapult.cpp
@@ -48,7 +48,7 @@ double Catapult::GetPosition()
 	return m_shooterController->GetPosition() * M_PI * 0.965;
 }
 
-double Catapult::GetPIDError()
+int Catapult::GetPIDError()
 {
 	return m_shooterController->GetClosedLoopError();
 }
